Update README.md
Hello, I'm Reagan Shannon, and this repository showcases my technical skills using the Twine website. The story I've crafted delves into Tudor history, offering a cause-and-effect narrative that immerses the player in King Henry's reign from the perspective of a poor shoemaker. The story is enhanced with graphics and interactive elements, allowing players to engage deeply in the narrative. I've also ensured that the text is 508 compliant for accessibility. If you encounter a dead end, you have the option to backtrack and choose a different path, potentially leading to the game's conclusion. Overall, I believe this game is a strong testament to my technical skills, and I eagerly await your feedback.